pyproject.toml
[project]
name = "mitt-vector-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "googledrivedownloader==1.1.0",
    "pytelegrambotapi==4.30.0",
    "python-dotenv==1.2.1",
    "segno==1.6.6",
]

[project.scripts]
bot = "bot.__main__:main"

[tool.uv]
package = true